fix: resolve all clippy -D warnings failures
lib/src/store.rs:
- redundant_closure: .map_err(|e| Error::Redis(e)) -> .map_err(Error::Redis)
- async_fn_in_trait: desugar all 5 GeoStore methods to RPITIT + Send bounds

lib/src/trie.rs:
- explicit_auto_deref: remove redundant &mut **child

demo/geo-node/src/main.rs:
- redundant_field_names: split_point: split_point -> split_point
- unnecessary_cast: entity_ttl_secs as u64 (already u64)

demo/geo-node/src/grpc.rs:
- unnecessary_cast: entity_ttl_secs as u64

demo/cluster-test/src/main.rs:
- empty_line_after_doc_comments (2 locations)
- doc_overindented_list_items (3 locations)
- cloned_ref_to_slice_refs: &[token.clone()] -> std::slice::from_ref

demo/loadtest/src/main.rs:
- too_many_arguments: #[allow] on sharded_writer_task
- len_zero (2): .len() > 0 -> !.is_empty()

demo/weather-server/src/{aggregate,metar_bulk}.rs:
- empty_line_after_doc_comments

demo/weather-server/src/routes.rs:
- dead_code: #[allow(dead_code)] on zoom field

+++ demo/geo-node/src/grpc.rs
@@ -85,7 +85,7 @@ pub struct GeoRedisService { pub state: crate::AppState }
 impl GeoRedisGrpc for GeoRedisService {
     async fn insert(&self, req: Request<GrpcGeoEntry>) -> Result<Response<InsertResponse>, Status> {
         let e   = req.into_inner();
-        let ttl = self.state.cfg.entity_ttl_secs as u64;
+        let ttl = self.state.cfg.entity_ttl_secs;
         let s2l = self.state.cfg.s2_level;
         let geo = GeoEntry {
             id: e.id.clone(), lat: e.lat, lon: e.lon,
